Add files via upload
I have used the column “Tweet” in NBC-FOX.csv dataset to create a clustering algorithm. Based on the structure of the data I analyse it using an appropriate type of clustering and also model input arguments. I use plotting tools to display the tweets in each cluster, and I  have used a classification model to classify tweets based on the news agency. Then I compare the resulted clusters with values in other columns in this dataset, such as "News agency" and "day."